requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
pytesseract==0.3.11
Pillow==9.6.1
python-dotenv==2.1.0
